desafios do projeto/desafio3_solution/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
argon2-cffi
python-multipart
